third_party/blink/web_tests/external/wpt/css/css-transitions/starting-style-cascade-expected.txt

This is a testharness.js-based test.
[FAIL] Starting style inheriting from parent's after-change style while parent transitioning
  assert_equals: Transition started from parent's after-change style color expected "rgb(0, 192, 0)" but got "rgb(0, 160, 0)"
[FAIL] Starting style inheriting from parent's after-change style while ancestor is transitioning
  assert_equals: Transition started from parent's after-change style color, inherited from ancestors after-change color expected "rgb(0, 192, 0)" but got "rgb(0, 160, 0)"
Harness: the test ran to completion.
